api_extract_info/requirements.txt
tensorflow
keras
torch
torchvision
pymongo
opencv-python
flask
flask-cors
Pillow
scipy
scikit-learn
h5py
pypiwin32